Implement Bayesian updates from Python script (#3)
* Start on posterior calculations, multiply, and normalize

* Make everything static in probability system

* Explicit cast from float to int in UnitStats for now...

* Implement the rest of bayesian updates

+++ Assets/Scripts/UnitStats.cs
@@ -100,7 +100,7 @@ public class UnitStats : MonoBehaviour, IComparable
 
     public void recieveDamage(float damage)
     {
-        this.health -= damage;
+        this.health -= (int)damage;
         Debug.Log("Recieved " + damage + " damage.");
         if (this.health <= 0)
         {
